+++ Kyoo.Common/Models/Studio.cs
@@ -5,7 +5,7 @@ namespace Kyoo.Models
 {
 	public class Studio
 	{
-		[JsonIgnore] public long ID { get; set; }
+		[JsonIgnore] public int ID { get; set; }
 		public string Slug { get; set; }
 		public string Name { get; set; }
